feat: logout button functionality refactor: change topbar to typescript
